+++ src/assets/style/components/_all.scss
@@ -2,6 +2,7 @@
 @import './event-update';
 @import './hero';
 @import './group';
+@import './group-banner';
 @import './modal';
 @import './roster';
 @import './scoreboard';
