+++ src/assets/style/components/_hero.scss
@@ -28,8 +28,12 @@
 
   &__text {
     font-family: $font-primary;
-    font-size: 5rem;
+    font-size: 3.5rem;
     font-weight: $font-family-lightest;
     color: $color-white;
+
+    @include breakpoint-sm {
+      font-size: 5rem;
+    }
   }
 }
