+++ src/assets/style/components/_roster.scss
@@ -8,7 +8,6 @@
   &__header {
     width: 100%;
     max-width: 650px;
-    // position: sticky;
     top: 0;
     display: flex;
     flex-direction: row;
